pyproject.toml
[project]
name = "chen"
version = "0.1.0"
description = "chen"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "httpx>=0.28.1",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot2[fastapi]>=2.4.3",
    "requests>=2.32.5",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_apscheduler"]
plugin_dirs = ["src/plugins"]
builtin_plugins = []
